Feat: #9 - Observable Pattern 적용 및 MVVM 리팩토링 완료
