+++ TVING/TVING/Resource/Observable/Observable.swift
@@ -1,0 +1,30 @@
+//
+//  Observable.swift
+//  TVING
+//
+//  Created by YOUJIM on 7/4/24.
+//
+
+import Foundation
+
+
+class Observable<T> {
+    typealias Listener = (T) -> Void
+    
+    var listener: Listener?
+    
+    var value: T {
+        didSet {
+            listener?(value)
+        }
+    }
+    
+    init(value: T) {
+        self.value = value
+    }
+    
+    func bind(_ closure: @escaping (T) -> Void) {
+        self.listener = closure
+        listener?(value)
+    }
+}
