+++ TVING/TVING/Source/ViewModel/LoginViewModel.swift
@@ -9,55 +9,49 @@ import UIKit
 
 
 class LoginViewModel {
-    var id: String = ""
-    var password: String = ""
+    var id = Observable(value: "")
+    var password = Observable(value: "")
+    var isIdTextFieldEmpty = Observable(value: true)
+    var isPasswordTextFieldEmpty = Observable(value: true)
     
-    func updateFieldValue(_ textField: UITextField) {
-        if textField.placeholder == "아이디" {
-            id = textField.text ?? ""
+    
+    init() {
+        setValues()
+    }
+    
+    func setValues() {
+        id.bind { id in
+            self.updateIdTextFieldStatus(id: id)
+        }
+        password.bind { password in
+            self.updatePasswordTextFieldStatus(password: password)
+        }
+    }
+    
+    func updateIdTextFieldStatus(id: String) {
+        if id == "" {
+            self.isIdTextFieldEmpty.value = true
         }
         else {
-            password = textField.text ?? ""
+            self.isIdTextFieldEmpty.value = false
         }
     }
     
-    func clearValue(_ textField: UITextField) {
-        textField.text = ""
-    }
-    
-    func updateLoginButtonStatus(_ loginButton: UIButton) {
-        if id.isEmpty || password.isEmpty {
-            loginButton.isEnabled = false
-            loginButton.backgroundColor = .black
-            loginButton.layer.borderColor = UIColor.gray04.cgColor
+    func updatePasswordTextFieldStatus(password: String) {
+        if password == "" {
+            self.isPasswordTextFieldEmpty.value = true
         }
         else {
-            loginButton.isEnabled = true
-            loginButton.backgroundColor = .accent
-            loginButton.layer.borderColor = UIColor.clear.cgColor
+            self.isPasswordTextFieldEmpty.value = false
         }
     }
     
-    func updateMask(_ maskButton: UIButton, _ passwordTextField: UITextField) {
-        passwordTextField.isSecureTextEntry.toggle()
-        if passwordTextField.isSecureTextEntry { maskButton.setImage(.icEyeOff, for: .normal) }
-        else { maskButton.setImage(.icEyeOn, for: .normal) }
-    }
-    
-    func pushToWelcomeVC(_ viewController: UIViewController) {
-        let welcomeVC = WelcomeViewController()
-        
-        // WelcomeVC에서 NotificationCenter로 id 데이터 받을 옵저버 등록
-        welcomeVC.receivedData()
-        
-        // NotificationCenter로 id 데이터 post
-        NotificationCenter.default.post(
-            name: NSNotification.Name("id"),
-            object: id
-        )
-        
-        welcomeVC.modalPresentationStyle = .fullScreen
-        
-        viewController.present(welcomeVC, animated: true)
+    func updateLoginButtonStatus() -> Bool {
+        if id.value.isEmpty || password.value.isEmpty {
+            return true
+        }
+        else {
+            return false
+        }
     }
 }
